Connection classes into process.connection package moved (aurusov/raox#334)
