+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/connection/ConnectionCreateCommand.java
@@ -6,19 +6,18 @@ import ru.bmstu.rk9.rao.ui.process.node.BlockNode;
 
 public class ConnectionCreateCommand extends Command {
 
+	private Connection connection;
 	private BlockNode sourceBlockNode;
 	private BlockNode targetBlockNode;
-	private Connection connection;
-
 	private String sourceDockName;
 	private String targetDockName;
 
-	public void setSource(BlockNode sourceBlockNode, String sourceDockName) {
+	protected final void setSource(BlockNode sourceBlockNode, String sourceDockName) {
 		this.sourceBlockNode = sourceBlockNode;
 		this.sourceDockName = sourceDockName;
 	}
 
-	public void setTarget(BlockNode targetBlockNode, String targetDockName) {
+	protected final void setTarget(BlockNode targetBlockNode, String targetDockName) {
 		this.targetBlockNode = targetBlockNode;
 		this.targetDockName = targetDockName;
 	}
@@ -50,13 +49,5 @@ public class ConnectionCreateCommand extends Command {
 	@Override
 	public void undo() {
 		connection.disconnect();
-	}
-
-	public final String getSourceDockName() {
-		return sourceDockName;
-	}
-
-	public final String getTargetDockName() {
-		return targetDockName;
 	}
 }
