+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/connection/ConnectionDeleteCommand.java
@@ -6,7 +6,7 @@ public class ConnectionDeleteCommand extends Command {
 
 	private Connection connection;
 
-	public void setConnection(Object model) {
+	protected final void setConnection(Object model) {
 		this.connection = (Connection) model;
 	}
