+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/connection/ConnectionDeleteEditPolicy.java
@@ -1,10 +1,8 @@
-package ru.bmstu.rk9.rao.ui.process.node;
+package ru.bmstu.rk9.rao.ui.process.connection;
 
 import org.eclipse.gef.commands.Command;
 import org.eclipse.gef.editpolicies.ConnectionEditPolicy;
 import org.eclipse.gef.requests.GroupRequest;
-
-import ru.bmstu.rk9.rao.ui.process.connection.ConnectionDeleteCommand;
 
 public class ConnectionDeleteEditPolicy extends ConnectionEditPolicy {
