+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/selectpath/SelectPathFigure.java
@@ -7,8 +7,8 @@ import org.eclipse.draw2d.IFigure;
 import org.eclipse.draw2d.geometry.PointList;
 import org.eclipse.draw2d.geometry.Rectangle;
 
+import ru.bmstu.rk9.rao.ui.process.connection.ConnectionAnchor;
 import ru.bmstu.rk9.rao.ui.process.node.BlockFigure;
-import ru.bmstu.rk9.rao.ui.process.node.ConnectionAnchor;
 
 public class SelectPathFigure extends BlockFigure {
